+++ src/features/imagePick/imagePickSlice.js
@@ -1,3 +1,4 @@
+import { createSlice } from "@reduxjs/toolkit";
 import { getActiveRoomId } from "../../selectors/selectors";
 import { updateText } from "../input/chatInputReducer";
 
@@ -15,6 +16,22 @@ export const selectImage = image => (dispatch, getState) => {
 	dispatch(updateText(activeRoomId, `${message} ${image} `));
 	dispatch(closeImageSelections());
 };
+
+const imagePickSlice = createSlice({
+	name: "imagePick",
+	initialState: {},
+	reducers: {
+		"images/received": (state, action) => ({
+			message: action.message,
+			images: action.images
+		}),
+		"images/close": () => null
+	}
+});
+
+// export const {  } = imagePickSlice.actions;
+
+export default imagePickSlice.reducer;
 
 // const testImages = [
 // 	"http://chariotlearning.com/wp-content/uploads/2015/12/Testing_in_Progress.gif",
